requirements.txt
boto3>=1.34.0
botocore>=1.34.0
psycopg2-binary>=2.9.0
requests>=2.31.0
sshtunnel>=0.4.0
pyyaml>=6.0
